requirements.txt
gitpython
javalang
pandas
requests
tqdm
